Add Frequency, Voltage and Serial to DEV
