+++ driver-gridseed.h
@@ -36,4 +36,19 @@
 
 extern struct device_drv gridseed_drv;
 
+typedef struct s_gridseed_info {
+	enum sub_ident	ident;
+	uint32_t	fw_version;
+	struct timeval	scanhash_time;
+	int		nonce_count[GRIDSEED_MAX_CHIPS];  // per chip
+	int		error_count[GRIDSEED_MAX_CHIPS];  // per chip
+	// options
+	int		baud;
+	int		freq;
+	unsigned char	freq_cmd[8];
+	int		chips; //chips per module
+	int		voltage;
+	int		per_chip_stats;
+} GRIDSEED_INFO;
+
 #endif /* INCLUDE_DRIVER_GRIDSEED_H */
